@liboz 2017-01-18
I reran the benchmark with the latest commit, which uses an `OptimizedClosure` even for lists of size 1, which makes quite a big difference for x86. I also included a list of size 200.

|         Method |       Jit | Platform | count |          Mean |     StdDev |        Median |  Gen 0 | Allocated |     Gen 0 |
|--------------- |---------- |--------- |------ |-------------- |----------- |-------------- |------- |---------- |---------- |
| mapFoldLibrary | LegacyJit |      X86 |     0 |     **8.4156 ns** |  0.2867 ns |     8.3025 ns | 0.0086 |      28 B |     20.17 |
|            old | LegacyJit |      X86 |     0 |    10.5356 ns |  0.0431 ns |    10.5390 ns | 0.0087 |      28 B |     18.37 |
| mapFoldLibrary |    RyuJit |      X64 |     0 |     **8.9394 ns** |  0.0521 ns |     8.9444 ns | 0.0175 |      56 B |     34.45 |
|            old |    RyuJit |      X64 |     0 |    11.3318 ns |  0.0449 ns |    11.3378 ns | 0.0175 |      56 B |     37.61 |
| mapFoldLibrary | LegacyJit |      X86 |     1 |    **20.1550 ns** |  0.0729 ns |    20.1465 ns | 0.0185 |      60 B |     36.50 |
|            old | LegacyJit |      X86 |     1 |    37.5498 ns |  0.1217 ns |    37.5273 ns | 0.0185 |      60 B |     35.28 |
| mapFoldLibrary |    RyuJit |      X64 |     1 |    **22.6510 ns** |  0.0547 ns |    22.6488 ns | 0.0351 |     112 B |     73.50 |
|            old |    RyuJit |      X64 |     1 |    25.0650 ns |  0.0735 ns |    25.0641 ns | 0.0350 |     112 B |     81.65 |
| mapFoldLibrary | LegacyJit |      X86 |     2 |    **47.1284 ns** |  0.3933 ns |    47.2077 ns | 0.0281 |      92 B |     55.07 |
|            old | LegacyJit |      X86 |     2 |    74.3761 ns |  0.1504 ns |    74.3878 ns | 0.0384 |     124 B |     73.24 |
| mapFoldLibrary |    RyuJit |      X64 |     2 |    **37.9762 ns** |  0.1207 ns |    37.9530 ns | 0.0528 |     168 B |    103.35 |
|            old |    RyuJit |      X64 |     2 |    48.4527 ns |  0.1928 ns |    48.4350 ns | 0.0727 |     232 B |    142.39 |
| mapFoldLibrary | LegacyJit |      X86 |     3 |   **55.9078 ns** |  0.7202 ns |    55.7897 ns | 0.0385 |     124 B |     81.38 |
|            old | LegacyJit |      X86 |     3 |   103.7128 ns |  0.5762 ns |   103.5408 ns | 0.0526 |     172 B |    114.98 |
| mapFoldLibrary |    RyuJit |      X64 |     3 |    **47.8466 ns** |  0.4055 ns |    47.7283 ns | 0.0702 |     224 B |    143.80 |
|            old |    RyuJit |      X64 |     3 |    65.6070 ns |  0.6674 ns |    65.8029 ns | 0.1007 |     320 B |    204.67 |
| mapFoldLibrary | LegacyJit |      X86 |     4 |    **64.7824 ns** |  0.6187 ns |    64.5236 ns | 0.0484 |     156 B |     93.75 |
|            old | LegacyJit |      X86 |     4 |   132.1718 ns |  1.2091 ns |   132.4878 ns | 0.0679 |     220 B |    135.02 |
| mapFoldLibrary |    RyuJit |      X64 |     4 |    **57.9229 ns** |  0.5386 ns |    57.7049 ns | 0.0880 |     280 B |    171.88 |
|            old |    RyuJit |      X64 |     4 |    84.8074 ns |  0.2946 ns |    84.7772 ns | 0.1285 |     408 B |    241.08 |
| mapFoldLibrary | LegacyJit |      X86 |     5 |    **73.1281 ns** |  0.2990 ns |    73.0463 ns | 0.0587 |     188 B |    115.19 |
|            old | LegacyJit |      X86 |     5 |   162.0976 ns |  1.6367 ns |   161.7521 ns | 0.0831 |     268 B |    168.11 |
| mapFoldLibrary |    RyuJit |      X64 |     5 |    **67.3089 ns** |  0.2145 ns |    67.3206 ns | 0.1056 |     336 B |    223.13 |
|            old |    RyuJit |      X64 |     5 |   102.7672 ns |  0.7567 ns |   102.9256 ns | 0.1555 |     496 B |    310.49 |
| mapFoldLibrary | LegacyJit |      X86 |    10 |   **117.1620 ns** |  1.4002 ns |   116.7370 ns | 0.1084 |     348 B |    222.90 |
|            old | LegacyJit |      X86 |    10 |   314.5876 ns |  2.6753 ns |   315.1066 ns | 0.1578 |     508 B |    333.40 |
| mapFoldLibrary |    RyuJit |      X64 |    10 |   **117.6482 ns** |  0.3742 ns |   117.6242 ns | 0.1937 |     616 B |    384.30 |
|            old |    RyuJit |      X64 |    10 |   201.4086 ns |  0.5629 ns |   201.4625 ns | 0.2932 |     936 B |    552.23 |
| mapFoldLibrary | LegacyJit |      X86 |   200 | **1,789.0379 ns** |  6.4853 ns | 1,789.4405 ns | 2.0108 |   6.43 kB |  4,174.38 |
|            old | LegacyJit |      X86 |   200 | 5,941.7822 ns | 49.1143 ns | 5,953.1631 ns | 2.9492 |   9.63 kB |  5,797.00 |
| mapFoldLibrary |    RyuJit |      X64 |   200 | **1,996.1305 ns** |  6.7861 ns | 1,997.3284 ns | 3.5443 |  11.26 kB |  7,476.49 |
|            old |    RyuJit |      X64 |   200 | 3,747.8527 ns | 14.4366 ns | 3,749.0509 ns | 5.5487 |  17.66 kB | 11,528.85 |