@dsyme 2017-01-18
@liboz I added some perf results here https://github.com/dsyme/visualfsharp/blob/2c3cece77c1287c5c0464d6f59759700d3cf4e70/tests/scripts/compiler-perf-results.txt

Some of the numbers look erratic - but overall I see only very small differences at the most for commit ``38e74ec89ada4bdd277be989a8aecfd855d1aaff ``, e.g. 30.4 instead of 30.9 for ``parseonly-to-check``.  Certainly there's no harm done but I think that ``mapFold`` doesn't dominate overall time