@KevinRansom 2017-02-24
@liboz @dsyme 

Do you have any recommendations for this PR?

Kevin